Add Player to Game class
