+++ src/game.js
@@ -1,4 +1,5 @@
 const fs = require('fs');
+const Player = require('../src/player.js');
 const Role = require('../src/role.js');
 module.exports = class Game{
     constructor(creator, createdOn, players){
@@ -32,4 +33,38 @@ module.exports = class Game{
         return roleList;
     }
 
+    assignPlayers(mentions, roles){
+		let list = [];
+		let person = null;
+        let role = null;
+		let index = null;
+
+		players.forEach(function(player){
+            person = player;
+            role = roles[Math.floor(Math.random() * roles.length)];
+            list.push(new Player(person, role));
+			index = roles.indexOf(role);
+			if (index !== -1) {
+				roles.splice(index, 1);
+			}
+		});
+		// return list of assign Players
+		return list;
+	}
+
+    messagePlayers(){
+		players.forEach(player => {
+			let message = new MessageEmbed();
+			message.setColor(Role.colors[player.Role.Color]);
+			message.setDescription(player.Role.Description);
+			message.setTitle(player.Role.Title);
+			// message.setThumbnail(player.Role.Image);
+			// message.setImage(player.Role.Image);
+			// p.person.send(embed=embedVar, file=file)
+			console.log(message)
+			//player.Person.send(message)
+
+		})
+		
+	}
 }
